+++ source/includes/integration/finite_elements/FiniteElementPolicy.h
@@ -5,10 +5,10 @@
 
 namespace csmp {
 
-template<size_t> class Node;
+template<uint32_t> class Node;
 
 /// finite element policy for class Element
-template<size_t dim, template<size_t> class CELL>
+template<uint32_t dim, template<uint32_t> class CELL>
 class FiniteElementPolicy {
   public:
     FiniteElementPolicy( FiniteElement* = nullptr );
@@ -19,10 +19,10 @@ class FiniteElementPolicy {
     void AssignFiniteElementNullPtr() { fptr_ = nullptr; }
     
     /// returns search key to match element faces; pointers in order so that they can be searched
-    std::set<Node<dim>*> CornerNodesOfFace( size_t face_id );
+    std::set<Node<dim>*> CornerNodesOfFace( uint32_t face_id );
 
     /// returns set of nodes that are neighbors of the target node in this element
-    std::set<Node<dim>*> CornerNodesConnectedTo( size_t node_id );
+    std::set<Node<dim>*> CornerNodesConnectedTo( uint32_t node_id );
 
     /// the type is an enumeration that is used in the generation of finite elements
     CSMP_FEM_TYPE  FE_Type() const;
@@ -37,15 +37,15 @@ class FiniteElementPolicy {
     bool       IsSurfaceElement() const;
     bool       IsVolumeElement() const;
 
-    size_t     Interpolation() const;
+    uint32_t     Interpolation() const;
     bool       UsesLocalCoordinates() const;
 
     /// number of element integration points for current quadrature scheme
-    size_t     IntegrationPoints() const;
+    uint32_t     IntegrationPoints() const;
   
     /// returns the location of the integration point in global coordinates
-    Point<dim> IntegrationPoint( size_t ip ) const;
-    double     WeightAtIntegrationPoint( size_t i ) const;
+    Point<dim> IntegrationPoint( uint32_t ip ) const;
+    double     WeightAtIntegrationPoint( uint32_t i ) const;
 
     /// mapping of local points from local to global coordinates
     Point<dim> RstToXYZ( Point<dim> rst ) const;
@@ -58,7 +58,7 @@ class FiniteElementPolicy {
     void    N_AtGlobalPoint( std::vector<double>& N, const std::vector<double>& xyz ) const;
     
     void    N_AtBaryCenter( std::vector<double>& N ) const;
-    void    N_AtIntegrationPoint( size_t ipoint, std::vector<double>& N ) const;
+    void    N_AtIntegrationPoint( uint32_t ipoint, std::vector<double>& N ) const;
 
     /// first (constant) derivatives of linear interpolation functions of analytically integrated simplex
     void	  Integral_dNT_K_dN(DenseMatrix<DM_MIN>& M, DenseMatrix<DM_MIN>& K) const;
@@ -66,16 +66,16 @@ class FiniteElementPolicy {
 	  void    dN( DenseMatrix<DM_MIN>& ) const;
   
     /// first derivatives of interpolation functions at the given node (returned is determinant of Jacobian matrix at i,j)
-    double   dN_AtNode( DenseMatrix<DM_MIN>&, size_t nd, size_t dof=1 ) const;
+    double   dN_AtNode( DenseMatrix<DM_MIN>&, uint32_t nd, uint32_t dof=1 ) const;
   
     /// first derivatives of interpolation functions at the (returned is determinant of Jacobian matrix at i,j)
-    double   dN_AtBaryCenter( DenseMatrix<DM_MIN>&, size_t dof=1 ) const;
+    double   dN_AtBaryCenter( DenseMatrix<DM_MIN>&, uint32_t dof=1 ) const;
   
     /// first derivatives of interpolation functions at Gauss point gp (returned is determinant of Jacobian matrix at i,j)
-    double   dN_AtIntegrationPoint( DenseMatrix<DM_MIN>&, size_t gp, size_t dof=1 ) const;
+    double   dN_AtIntegrationPoint( DenseMatrix<DM_MIN>&, uint32_t gp, uint32_t dof=1 ) const;
 
     /// returned is determinant of Jacobian matrix at given Gauss point
-    double   det_JINV_AtIntegrationPoint( size_t ipoint ) const;
+    double   det_JINV_AtIntegrationPoint( uint32_t ipoint ) const;
 
     /// interpolation function products matrix for analytically integrated element
     void     IntegralNN( DenseMatrix<DM_MIN>& M ) const;
@@ -102,17 +102,17 @@ class FiniteElementPolicy {
   
     /// returns the value of any property interpolated to the integration point of interest
     template<class Var>
-    void       PropertyValueAtIntegrationPoint( const csmp::Index&, size_t integration_point, Var& ) const;
+    void       PropertyValueAtIntegrationPoint( const csmp::Index&, uint32_t integration_point, Var& ) const;
 
     /// returns the value of a scalar property interpolated to the integration point of interest
-    double   PropertyValueAtIntegrationPoint( const csmp::Index&, size_t integration_point ) const;
+    double   PropertyValueAtIntegrationPoint( const csmp::Index&, uint32_t integration_point ) const;
 
     /// returns property values at the integration points
     template<class Var>
     void       IntegrationPointPropertyVector( const csmp::Index&, std::vector<Var>& ) const;
 
     /// uses linear extrapolation of property values from the integration points to the nodes
-    void       ExtrapolateIntegrationPointVariableToNodes( size_t nvars,
+    void       ExtrapolateIntegrationPointVariableToNodes( uint32_t nvars,
                                                            const std::vector<double>&integration_point_var,
                                                            std::vector<double>& node_variable ) const;
 
@@ -126,31 +126,31 @@ class FiniteElementPolicy {
     double   InnerRadius() const;
   
     /// number of edges = segments in the finite element of interest
-    size_t     Segments() const;
+    uint32_t   Segments() const;
 
     /// retrieve position of the segment (element edge) mid-points
-    Point<dim> SegmentMidPoint( size_t segm ) const;
+    Point<dim> SegmentMidPoint( uint32_t segm ) const;
 
     /// returns the length of the nth segment=edge of the finite element
-    double     SegmentLength( size_t segm ) const;
+    double     SegmentLength( uint32_t segm ) const;
   
     /// returns a vector with the lengths of all finite element segments = edges
     void       SegmentLengths( std::vector<double>& ) const;
 
     /// retrieve barycenter of element face
-    Point<dim> FaceBaryCenter( size_t face ) const;
+    Point<dim> FaceBaryCenter( uint32_t face ) const;
 
     /// returns area of face i (to be scaled with thickness attribute if this is a lower-dimensional element)
-    double     FaceArea( size_t face ) const;
+    double     FaceArea( uint32_t face ) const;
 
     /// returns unit normal to face of element i
-    Point<dim> UnitNormalToFace( size_t face ) const;
+    Point<dim> UnitNormalToFace( uint32_t face ) const;
 
     /// returns unit normal to face of element i into STL vector; fastest version!
-    void       UnitNormalToFace( size_t face, std::vector<double>& ) const;
+    void       UnitNormalToFace( uint32_t face, std::vector<double>& ) const;
   
     /// returns unit normal to face of element i into supplied csmp VectorVariable
-    void       UnitNormalToFace( size_t face, VectorVariable<dim>& ) const;
+    void       UnitNormalToFace( uint32_t face, VectorVariable<dim>& ) const;
   
     /// computes unit normal to element if it is planar (3D) or linear (2D); vector components are returned into point
     Point<dim> UnitNormal() const;
